Delete album when all of its tracks have been deleted
When an entire album or the last track from an album is deleted the
entry in the "albums" table is not removed. If tracks from the same
album are later found again by the MediaScanner then the album will
not get any album art.
Fixed by adding a trigger that removes the "albums" table entry if
no tracks with the same album_id are found after a delete.

Change-Id: I31d30bfe9cf067e6378b3bb3c4f53152b9c1e5cf