Notify change of movePlaylistEntry after transaction ends
When a client asks the Media Provider to rearrange a playlist
the notifyChange callback function is called before the database
transaction is ended resulting in that listeners can't get the
new data when they get the callback.

notifyChange() is called before the database transaction is ended
resulting in that listeners can't get the new data when they get the
callback.

This fix will wait until the database transaction is finished
to do the callback.

Change-Id: Ifdfbb047b809bd5fba4752bcdd299927f9082eee