Update patch set 1
Patch Set 1:

Just to be clear, this patch solves a problem we observe on devices with removable media (external SD slots), so the Nexus S (4G) and Galaxy Nexus are unaffected.

This patch, and the one dependent on it (I7e307225) we've been using in CM9 and CM7 for a couple of weeks now with much success.

Here's a write-up that describes the "disappearing SD media" issue in more detail:
http://epiccm.blogspot.com/2012/03/details-on-disappearing-sd-media-issue.html

Patch-set: 1
Label: Code-Review=0