The AlbumArt image of songs in the non-primary storage root director are those images in the same direcotry
MediaProvider will fall back to the image in the same folder of music,
if MetadataRetriver failed to extract the AlbumArt image from those mp3 files.
MediaProvider will only bypass the albumArt image search if the music located
in the root director of the primary external storage. Add this bypass for all
external storages.

Change-Id: I972586b2f92bd439f51ee932f09e824fc5507236
Signed-off-by: guoyin.chen <guoyin.chen@freescale.com>