Primary storage of multiple users in MediaProvier and MTP
It is assumed that a primary storage for each user is the first entry of
storage list managed by MediaProvider and MTP (espeically in getStorageId()).
But in a multiple users environment, a primary storage for a user
(e.g. /storage/emulated/10) is the last entry of the list because the new
storage is appended to the existing volume list of MountService after the new
user is created.

This patch makes sure MediaProvider finds a primary storage first and then adds
the other storages so that files MediaProvider database have correct storage_id.

Change-Id: Ib1dad5b00f0c71de601e57a47aa7191f593a94a6