Update patch set 1
Patch Set 1:

Note that the datetime() *function* has second granularity, however the code uses the datetime *type* not the function.
The values stored in the field actually have millisecond granularity, because they are generated with "strftime('%Y-%m-%d %H:%M:%f','now')". That last "%f" gives fractional seconds, down to the millisecond.

There is of course still a small risk of duplicates even at the millisecond level, but in our internal tree we chose to simply have the logToDb() method do an "insert or replace" instead of an "insert" to solve that.

If you still want to make this change, then you should bump the schema version and add an update step that recreates this table with the new schema, preserving previous data. As currently written, this change won't work for devices that are updated.

Patch-set: 1
Label: Verified=0