When deleting a database file, it should confirm if the file is a real database file. In case of media provider, it uses Write Ahead Logging mode and there could be wal and shm files for databases. After media scanning, basically 6 files can exist, 3 for internal db, 3 for external db. In that case, unfortunately, a wal file or shm file of a valid database can be deleted during media provider initialization. (DatabaseHelper.onOpen) If a wal file is deleted, db can't guranteee that user data stays safe.
Of course, if "mContext.databaseList()" returns the list of only database files, this patch is not necessary.
But we can not guarantee that all database files have ".db" extension.
And we can easily see that some market apps doens't use ".db" as a database file extension.
So, I think that fixing media provider is right.

Change-Id: Ie5de421593e2d10aab8398006bc80b3e18b7c125