Update patch set 1
Patch Set 1: Code-Review-1

This seems like the wrong way to go about this. You're putting a limit on the compressed size of an image, without taking into account how much memory is already in use, and without taking into account the uncompressed size of the image.
With this change, the media provider will still crash if the compressed image is smaller than the calculated heap size, but larger than the free space, for example.
It would be better to just catch the OutOfMemoryError.

Patch-set: 1
Label: Code-Review=-1