Update patch set 1
Abandoned

Thanks Marco for your review. We recognized this patch didn't fully solve the issue as you pointed out.

What we wanted to achieve is to make DATA field up-to-date when the name of playlist changed. Currently, even though the user changes the name of playlist, DATA field of corresponding entry in DB is kept untouched. Thus, exception happens if user tries to create a playlist of which name is same as previously-created (but renamed after) playlist.


We'll modify the patch and come back.

Patch-set: 1
Status: abandoned