Update patch set 1
Patch Set 1:

Sorry for my poor English. I mean to say if app provide the delete option to delete something, there probably be a case to check the file is really deleted or not after reboot the device. That makes me think we can add the line in MediaProvider. From this point, the result is the same.
I fully understand your concern, some apps can do something else before the file really get deleted. It will depend on the possibility of such case.

Patch-set: 1