Fix getting wrong size of external SDcard via MTP
Issue description:
It gets wrong size of external SDcard via MTP when SDcard is inserted 
after connecting DUT to PC via USB cable.

Steps to reproduce the issue:
1. Connect DUT to Win7 PC via USB cable.
2. Insert an external SDcard into DUT.
3. Check availabe and total size of SDcard on PC.

Expected Results:
It shows the actual size of the external SDcard.

Actual Results:
It does not show the actual size of the external SDcard via MTP on PC.

Root cause:
There are two mount points for external SDcard, one is
"/mnt/media_rw/sdcard1" whose type is vfat, the other one is
"/storage/sdcard1" which is a fuse filesystem and it is the directory
that we can actually see via MTP. When external SDcard is mounted, it
just means that it's mounted to the point "/mnt/media_rw/sdcard1".
However the fuse "/storage/sdcard1" isn't ready until the whole SDcard
is scanned by MediaScannerService, not just by checking if the state of
the volume is MEDIA_MOUNTED.

Solution:
Don't send evet "MTP_EVENT_STORE_ADDED" until finishing scan directory
"/storage/sdcard1".

Change-Id: Ic0902a1bc6a2f0292ddf4a1ac07927a19c6e7885
Signed-off-by: Qin Chao <chaox.qin@intel.com>
Signed-off-by: Jackie Wu <jackie.wu@intel.com>