Update patch set 4
Patch Set 4:

Hi Macro, yes it's still accessible to "/storage/sdcard1", but the total size and available size of sdcard are not shown the actual szie   on PC via MTP, actually smaller than the real size. And this will cause to failure of copying file whose size is larger than the available size shown on PC, but there is enough free space for the file on sdcard. It cannot get the right size until media scanner is finished.

Patch-set: 4
Label: Verified=0