Update patch set 1
Patch Set 1:

Correct, I found that it works well in my testing. If the directory is not there then the ptp share doesn't show it on the windows side (or linux, or mac...). So you will see the ptp share, but it will not contain any folders.

When/if one of the folders gets created on the device; then a refresh on the host side will cause the folder(s) to show up with any files they may contain.

Patch-set: 1
Label: Verified=0