Update patch set 2
Patch Set 2:

So if the pictures folder doesn't exist anywhere, nothing will show on the host side? Does that mean you won't be able to copy *to* the device either in that case? If so I think that may be why the original code created the folder if it didn't exist, and we should probably keep that behavior at least for the primary.

Patch-set: 2
Reviewer: Marco Nelissen <1001525@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0