Add process of deletion for audio file and video thumbnail
This change can fix these issues.

1.
When deleting image and video content in MediaProvider,
file is deleted with information of DB.
But only information of DB is deleted about audio contents.
Therefore when a Settings application deletes audio content,
file is left.

2.
When deleting image file, thumbnail is deleted together.
But even if video file is deleted, thumbnail is left.

Change-Id: I074d2f6e615a2e22512036ad199914c3f9bc9222