Update patch set 1
Patch Set 1:

My concern with this change is that the media provider has never deleted audio files, that's always been up to the app, and so I worry about backward compatibility.
Perhaps it would be a good idea to make this dependent on the targeted API version of the caller?

Patch-set: 1