Update patch set 1
Patch Set 1:

In our understanding, delete functionality in Settings is new from M (MR1?). To invoke this issue, application should use MediaProvider. But it's maybe not common usecase before M. To delete actual file in previous version, application has to delete the file directly, via MTP or File Commander etc.


And we think it should not differentiate the behavior for each version. It complicates the code & behavior. We think this is a simple bug. Other kinds of files, video and image, deletes actual files. We don't think 3rd party application change the behavior in each kinds of files.
(On the other hand, I also understand this is quite simple, so it may affect widely. But I suggest to go forward with simple solution.)

Patch-set: 1