Update patch set 1
Patch Set 1:

Hi Jeff. Could you please review this? I'm not 100% sure if we can remove this.

Currently we have two complete scans on each power cycle. Although the second one does not really parse file, it still take a lot of time if we have many file. We made 20000 files on Pixel and ZTE Axon 7 and tested the time consumed on second scan. Here is the result:
Pixel: 2m3s
Axon 7: 1m46s
This is only the time spent on traversing the 20000 dummy files. They are not media files and no one is parsed.

We have many users who have been using a device for a long time and have 100000+ files on the storage. These are mostly junk files produced by all sorts apps. These users always complain their handsets get warmer on boot up for a long time and it turned out to be the scanner.

We have tested this change on our Axon 7 for a while. It seems nothing is broken. But we are not sure why the external scan is added in the first place.

Patch-set: 1
Reviewer: Alvin Cao <1073563@85c56323-6fa9-3386-8a01-6480fb634889>