Update patch set 1
Patch Set 1:

As long as MEDIA_MOUNTED is sent for the external storage it should be fine to remove the scan for BOOT_COMPLETED. I don't know if thats' guaranteed though.

Also, 2 minutes seems way too long to rescan 20k files, and a regression from what it used to be. I have an old doc in which I wrote down some measurements, and it lists 40 seconds for a rescan of 31000 files, and that was 3 years ago on a much slower device.

Patch-set: 1
Reviewer: Marco Nelissen <1001525@85c56323-6fa9-3386-8a01-6480fb634889>