Update patch set 1
Patch Set 1:

> As long as MEDIA_MOUNTED is sent for the external storage it should
 > be fine to remove the scan for BOOT_COMPLETED. I don't know if
 > thats' guaranteed though.
 > 
 > Also, 2 minutes seems way too long to rescan 20k files, and a
 > regression from what it used to be. I have an old doc in which I
 > wrote down some measurements, and it lists 40 seconds for a rescan
 > of 31000 files, and that was 3 years ago on a much slower device.

Marco, is it possible to find the author of this line and ask why this is added in the first place? The second scan also takes a lot of resources. I think most people agree that scanning on mount makes more sense than on boot complete and scanning only once is sufficient. But it seems no one is 100% sure. From our test on Axon 7, so far so good.
And here is how I tested the performance:
* Enable the following debug code in MediaScannerService.java:
  if (true) Log.d(TAG, "start scanning volume " + volume + ": "
                                + Arrays.toString(directories));
  scan(directories, volume);
  if (true) Log.d(TAG, "done scanning volume " + volume);
* Factory reset.
* Make dummy files.
 * cd /mnt/sdcard/Pictures; dd if=/dev/zero of=./dummy bs=1k count=1; for i in `seq 1 10000`; do cp dummy dummy$i; done
 * repeat above in /mnt/sdcard/Movies
* Clear the media storage database for Settings
* Reboot
* adb logcat -s MediaScannerService to get the log and calculate the time consumed from log timestamps.

Patch-set: 1
Reviewer: Alvin Cao <1073563@85c56323-6fa9-3386-8a01-6480fb634889>