Update patch set 3
Patch Set 3:

Marco, it was pointed out that this change is causing some issues.

I've been able to reproduce the issues and found the cause. When a query specifies the position of the whereArg in the array (e.g. ?1), duplicating the number of parameters makes SQLiteProgram throw an exception.

I don't know what's the best way to solve this problem. Maybe SQLiteProgram could be change so that it ignores the extra arguments.

In any case, I think it would be a good idea to show a warning message whenever the expected number of arguments differs from the number of arguments provided. That would make catching bugs such as the one I tried to fix with this commit easier.

Patch-set: 3
Reviewer: Gabriele M <1144723@85c56323-6fa9-3386-8a01-6480fb634889>