Handle null values in parent column of files table
If a null value is present in parent column, deleting entries from
database is not possible. This is caused by current logic which
prevents the deletion of an entry which is a parent for any other
entry in the database. Modification is needed to handle a case of
null value in parent column.

Test: put null in parent column and try to delete the file

Change-Id: Ic10ef6f63a9f2788ccb9ca2790c369ff09fa4c0c