fix MediaProvider ANR when connect MTP
When connect MTP ACTION_USB_STATE may receive by MtpReceiver, the function
MediaProvider insert will be called to tell MediaProvider MTP is configured.
In the function of insert, fixParentIdIfNeeded() will be called, which will
operate sqlite database. Then, main thread was blocked, which may result in
ANR.
This happens in thousands of users' device. And could be reproduced with
following method:
Make an directory under sdcard, such as /sdcard/test. Generate 10 thousand
files under the directory of test, then reboot the device to trigger scan of
the files. After scan, delete the directory of /sdcard/test, and trigger scan of
/sdcard/test with the command of am broadcast. Then, the database of test has been
remove from sqlit, but the files has not be move. Connect the device with MTP,
fixParentIdIfNeeded will be called, and ANR happened.

Test: as described above.

Change-Id: I818111c21c36a497eb275a73e206e099832c800d
Signed-off-by: zengjing <zengjing@xiaomi.com>