Support RRO for ringtones, notifications and alarms
Copy ringtone, alarm and notification from resources if an RRO has been
applied to the MediaProvider apk with the proper category.

Note that the default resources are empty binaries and a default sound
overlay will only be copied if its size is not zero.

Test: test default resources, install and enable
1. Make sure that resources are used when a newly flashed device
is booted.
2. Make sure that a newly installed and enabled test RRO with media
resources changes the resources used.

Bug: 69299998
Change-Id: I3bba5ad9e59c1f9aeb4af7a9266f073c49c2df74