Update patch set 3
Patch Set 3:

(2 comments)

These new columns seem very specific, and don't look like information that would normally be presented to the user or used for filtering or sorting. (based on our discussions I was expecting something like an "is_hdr" column so that an "HDR" badge could be shown on thumbnails, for example)
Is it actually important to have all three values in the database?

Patch-set: 3
Reviewer: Gerrit User 1001525 <1001525@85c56323-6fa9-3386-8a01-6480fb634889>