Update patch set 3
Patch Set 3:

(1 comment)

We want to avoid to use "is_hdr" since the definition of HDR is difficult and varies (HDR is relatively new technology, so new standards might be added in the near future. We would like to handle those cases as well without framework modification).

So, we would like to make a room for application to embed HDR detection logic in app side.

Patch-set: 3
Reviewer: Gerrit User 1233895 <1233895@85c56323-6fa9-3386-8a01-6480fb634889>