[MediaProvider] set usesCleartextTraffic prop
DownloadProvider share the same process name with MediaProvider. In DownloadProvider,
the prop of "usesCleartextTraffic" was set true, when MediaProvider does not set.
If MediaProvider was loaded first, FLAG_USES_CLEARTEXT_TRAFFIC will be set as default
for DownloadProvider (they share the same process name). In Android P, the default value
of this prop was set to false. Thus, everything download by DownloadProvider could
not be access by HTTP. Set this prop for MediaProvider will fix it.

Test: run cts-on-gsi -m CtsNetSecConfigDownloadManagerTestCases -t android.security.net.config.cts.DownloadManagerTest#testPerDomainCleartextAccepted

Change-Id: Ie7679cacc94a9bc913fe3b35d71d95e3d6a41a4b
Signed-off-by: zengjing <zengjing@xiaomi.com>