New queryRecentDocs in MediaDocumentsProvider
This implements the new DocumentsProvider#queryRecentDocuments API that
supports flexible limit for max number of recent documents returned.

Bug: 111288304
Test: Observed that 64 files are returned with old API, 64 files are
returned with the new API without extra args, and 100 files are returned
with the new API after passing it as extra args (with the
QUERY_ARG_LIMIT is included in the honored list).

Change-Id: Ia42e8aeac62379b35b85af848698eb3ece3f7947