Update patch set 5
Patch Set 5: -Autosubmit

> Patch Set 5:
> 
> > Patch Set 4:
> > 
> > (fwiw, i assume you'll need to remove the "DO NOT MERGE" here.)
> 
> Okay, thank you. I don't know why that was added to commit message.

@Elliott, @Anastasiia, this patch conflicts with the internal master branch as the code has diverged for future android versions. Therefore automerging will fail, the Presubmit failed too (Patchset 3). To stop the auto-merging, I added the "Do not merge" in the commit message. Without Presubmit-Verified +1, this CL will not be submitted.

Actually to think again, if the module owners review this change and agree this is a valid change (with a +1 or +2), this CL can then be modified and merged in the internal master. Since this CL will never be upstreamed at this state (Android 10 is released and Android 11 has a conflict), there is no point of merging this CL to aosp master. Partners would then include this CL in their Android 10 builds, but a version of this fix will be present in Android 11.

Patch-set: 5
Reviewer: Gerrit User 1377410 <1377410@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=0
Label: Presubmit-Ready=0