Fix media database upgrades failure
Delete MediaUpgradeReceiver, because PRE_BOOT_COMPLETED is too late
for media database to upgrade.

The function of updateDatabase does too many things and often results
in timeout publishing content providers.

So adding columns is keep in it, and most updating of content values
is moved to the function of doUpgrade.

Test: Database upgrades from P to Q
Change-Id: Ieecea02da3815833915ce910164b8f2a803972ab