DO NOT MERGE Enable fuse passthrough mode in FuseDaemon
FUSE passthrough is a feature that allows creating at OPEN/CREATE time a
direct channel from the process accessing the FUSE file system to the
file that lies in the lower file system. All the following read/write
operations targeting a file that was opened with FUSE passthrough mode
on target the lower file system directly, without the overhead of
traversing the FUSE daemon.

  go/fuse-passthrough

This patch enables the FUSE passthrough feature in the fuse daemon
spawned by MediaProvider.
FUSE passthrough is activated according to the

  persist.sys.fuse.passthrough.enable

flag. Being this patch for testing purpose, FUSE passthrough is enabled
by default and the flag can only be used to disable the feature is set
to false.
No further special conditions of the file are required for the
passthrough feature to be enabled.

Bug: 174548608
Test: ScopedStorageTest
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: If674367e5dcdfc7c71a262a7748be4a1b23eb47e