Don't modify rows inside recomputeMediaTypeValues iteration
recomputeMediaTypeValues iterated over all rows in the database where
media_type was not set, in order to find old entries that match the new
MEDIA_TYPE_DOCUMENT and MEDIA_TYPE_SUBTITLE media types. The rows were
updated in-loop to the new types.
But if the number of rows in the selection was greater than the first
CursorWindow, then the next window would re-run the query and start at
the wrong row, since modified rows of the first window may no longer
match the query. The Cursor would then step outside of the last window
leading to an IllegalStateException.

Since the operation performed is a find-and-update, replace the loop
with one update for each new media type instead.

Also fixes a small bug where existing files with text/vtt would be
marked as MEDIA_TYPE_DOCUMENT by recomputeMediaTypeValues.

Bug: 171659403
Test: Manual - see issue
Change-Id: I90cba6fe525c4c1b25235a3ed55affaeaeb3ddda